finish moving chat -> chat-playground / ui and updating the prompts
