+++ llama_stack/ui/app/chat-playground/page.tsx
@@ -151,7 +151,7 @@ const handleSubmitWithContent = async (content: string) => {
   }
 };
   const suggestions = [
-    "What is the weather in San Francisco?",
+    "Write a Python function that prints 'Hello, World!'",
     "Explain step-by-step how to solve this math problem: If x² + 6x + 9 = 25, what is x?",
     "Design a simple algorithm to find the longest palindrome in a string.",
   ];
