+++ llama_stack/ui/app/logs/vector-stores/page.tsx
@@ -16,8 +16,8 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@/components/chat/table";
-import { Skeleton } from "@/components/chat/skeleton";
+} from "@/components/ui/table";
+import { Skeleton } from "@/components/ui/skeleton";
 
 export default function VectorStoresPage() {
   const client = useAuthClient();
