+++ llama_stack/ui/components/layout/page-breadcrumb.tsx
@@ -9,7 +9,7 @@ import {
   BreadcrumbList,
   BreadcrumbPage,
   BreadcrumbSeparator,
-} from "@/components/chat/breadcrumb";
+} from "@/components/ui/breadcrumb";
 
 export interface BreadcrumbSegment {
   label: string;
