+++ llama_stack/ui/components/responses/items/web-search-item.tsx
@@ -1,7 +1,7 @@
 import {
   MessageBlock,
   ToolCallBlock,
-} from "@/components/chat/message-components";
+} from "@/components/chat-playground/message-components";
 import { WebSearchCallItem } from "../utils/item-types";
 
 interface WebSearchItemProps {
